e-opresional upgrade

+++ app/Http/Controllers/Ikm/Jadwal.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers\Ikm;
 
-use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
-use App\Models\Ikm\Jadwal as Model;
 use Carbon;
+use Illuminate\Http\Request;
+use App\Models\Ikm\Jadwal as Model;
+use App\Http\Controllers\Controller;
 
 class Jadwal extends Controller
 {
@@ -40,9 +40,9 @@ class Jadwal extends Controller
     {
         $request->validate([
 
-            'start_date' => 'required',
-            'end_date' => 'required',
-            'keterangan' => 'required|min:8'
+            'start_date' => 'required|unique:ikm',
+            'end_date' => 'required|unique:ikm',
+            'keterangan' => 'required|min:8|unique:ikm'
 
         ]);
 
@@ -61,13 +61,46 @@ class Jadwal extends Controller
         ->with('success', 'Data Berhasil Ditambah');
     }
 
+    public function show(Request $request, int $id)
+    {
+        $cek = Model::where('id', $id)->first();
+
+        if (strtotime($cek->end_date) < strtotime(date('Y-m-d'))) {
+            return redirect(route('intern.ikm.settingikm.index'))
+            ->with('warning', 'IKM ini sudah kadaluarsa');
+        }
+
+        $cek = Model::where('is_open', 1)->get();
+
+        if (count($cek) === 1 && $request->submit === 'Open') {
+
+            return redirect(route('intern.ikm.settingikm.index'))
+            ->with('warning', 'Hanya diperbolehkan 1 survey IKM saja yang aktif');
+        }
+
+       $ikm = Model::find($id);
+
+        if ($ikm->is_open === '') {
+
+            $request->is_open = NULL;
+
+        }
+
+        $ikm->is_open = $request->is_open;
+
+        $ikm->save();
+
+        return redirect(route('intern.ikm.settingikm.index'));
+ 
+    }
+
     /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(int $id)
     {
         $settingikm = Model::find($id);
 
@@ -81,7 +114,7 @@ class Jadwal extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id)
     {
         $request->validate([
 
@@ -111,8 +144,20 @@ class Jadwal extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(int $id)
     {
+        $jadwal = Model::find($id);
+
+        if ($jadwal->is_open === 1) {
+            return redirect(route('intern.ikm.settingikm.index'))
+            ->with('warning', 'Tidak diperbolehkan menghapus survey yang sedang berlangsung');
+        }
+
+        if(count($jadwal->result) !== 0){
+            return redirect(route('intern.ikm.settingikm.index'))
+            ->with('warning', 'IKM ini sudah terisi oleh beberapa responden dan tidak dapat dihapus');
+        }
+
         Model::destroy($id);
 
         return redirect(route('intern.ikm.settingikm.index'))
